-Simplified GridCell class by removing the hash, the Value enum and replacing all double values by float values. -replaced double values with float values and Point3d objects with Point3f objects in most classes to reduce memory usage. -Xwalk is now ~2-5x faster and requires 2-3 times less memory

+++ trunk/Xwalk/src/structure/math/Mathematics.java
@@ -225,7 +225,7 @@ public class Mathematics {
             minY = Math.min(minY, point.getY());
             minZ = Math.min(minZ, point.getZ());
         }
-        return new Point3d(minX, minY, minZ);
+        return new Point3d((float)minX, minY, minZ);
     }
     //--------------------------------------------------------------------------
     /**
@@ -448,13 +448,13 @@ public class Mathematics {
     /**
      * Calculates the Euclidean distance between two Cartesian coordinates.
      * @param  xyz1
-     *         Point3d object holding the first set of Cartesian coordinates.
+     *         Point3f object holding the first set of Cartesian coordinates.
      * @param  xyz2
-     *         Point3d object holding the second set of Cartesian coordinates.
+     *         Point3f object holding the second set of Cartesian coordinates.
      * @return Euclidean distance as a double value.
      */
-    public static double distance(final Point3d xyz1, final Point3d xyz2) {
-        return Math.sqrt(Math.pow(xyz2.getX() - xyz1.getX(), 2)
+    public static float distance(final Point3f xyz1, final Point3f xyz2) {
+        return (float)Math.sqrt(Math.pow(xyz2.getX() - xyz1.getX(), 2)
                        + Math.pow(xyz2.getY() - xyz1.getY(), 2)
                        + Math.pow(xyz2.getZ() - xyz1.getZ(), 2)
                         );
